Implements refresh token functionality
Adds refresh token functionality to the login process.

This change introduces refresh tokens to improve security and user experience.
It generates access and refresh tokens upon successful login. The refresh token is stored in a cookie.

A new endpoint is added to refresh access tokens using the refresh token stored in the cookie.